feat: refactor component with a texarea for Objectifs
